faiss-db/requirements.txt
fastapi
uvicorn[standard]
pydantic
faiss-cpu        
numpy            